requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.14.0
Werkzeug==3.1.3
python-dotenv==1.0.0
supabase==2.0.3
httpx==0.24.1
